pyproject.toml
[tool.poetry]
name = "wiredflow"
version = "0.1.0"
description = ""
authors = ["Dreamlone <mik_sar@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
attrs = "^22.2.0"
requests = "^2.28.2"
schedule = "^1.1.0"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
